\name{RVideoPoker-package}
\alias{RVideoPoker-package}
\alias{RVideoPoker}
\docType{package}
\title{Play Video Poker within R
}
\description{
Play Video Poker with R, complete with a graphical user interface. So
far, only "Jacks or Better" is implemented.
}
\details{
\tabular{ll}{
Package: \tab RVideoPoker\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-04\cr
License: \tab GPL-2\cr
}
}
\author{
  Roland Rau
  
  Maintainer: Roland Rau <roland.rau@gmail.com>
}
\keyword{ package }
\seealso{
See also my other game package: Miney
}
\examples{
cat("All meaningful examples would require a GUI. Hence this nonsense
example.\n Typically, you would enter\n rvideopoker()\n to start a
standard session.\n Most importantly: Have fun!\n")
}
